refactor(server): isolate authority adoption lifecycle
