+++ services/server/src/app.ts
@@ -120,9 +120,15 @@ import {
 import {
   registerConnectorInventoryRoutes
 } from "./features/connectors/inventory-routes.js";
+import {
+  registerAuthorityAdoptionRoutes
+} from "./features/authority-adoption/routes.js";
 import { sessionToken } from "./platform/session-cookies.js";
 import { audit } from "./platform/audit-events.js";
-import { authorityUrl } from "./platform/authority-url.js";
+import {
+  authorityImportCapability,
+  authorityUrl
+} from "./platform/authority-url.js";
 import {
   authenticatedUser,
   bearerToken,
@@ -224,27 +230,6 @@ interface AuthorityImportTransferRow {
   manifest_digest: string | null;
   source_revision: string | null;
   expires_at: string | Date;
-}
-
-interface AuthorityAdoptionRow {
-  id: string;
-  secret_hash: string;
-  collection_id: string;
-  display_name: string;
-  source_name: string;
-  retain_mirror: boolean;
-  mirror_name: string | null;
-  user_id: string | null;
-  state: "requested" | "approved" | "prepared" | "activating" | "completed" | "cancelled" | "expired";
-  next_authority_epoch: string | number;
-  final_head: string | number | null;
-  manifest_digest: string | null;
-  source_revision: string | null;
-  contracts: CollectionContractDescriptor[];
-  expires_at: string | Date;
-  approved_at: string | Date | null;
-  prepared_at: string | Date | null;
-  completed_at: string | Date | null;
 }
 
 interface AuthorityPairing {
@@ -430,513 +415,12 @@ export async function buildApp(options: BuildOptions) {
     tailscaleAuth: options.tailscaleAuth
   });
   registerConnectorInventoryRoutes(app, { db: options.db });
-
-  app.post("/v1/authority-adoptions", async (request, reply) => {
-    if (!options.hostedCollections || !options.hostedProvider) {
-      return reply.code(404).send(apiError(
-        "hosted_adoption_unavailable",
-        "This Connect server cannot adopt local collections."
-      ));
-    }
-    await recoverExpiredAuthorityAdoptions(options.db, options.hostedProvider);
-    const input = z.object({
-      collection_id: z.uuid(),
-      display_name: z.string().trim().min(1).max(200),
-      source_name: z.string().trim().min(1).max(200),
-      retain_mirror: z.boolean().default(true),
-      mirror_name: z.string().trim().min(1).max(200).optional()
-    }).strict().parse(request.body);
-    const id = randomUUID();
-    const secret = randomToken("adp");
-    const expiresIn = 30 * 60;
-    await options.db.query(
-      `INSERT INTO authority_adoption_requests
-         (id, secret_hash, collection_id, display_name, source_name,
-          retain_mirror, mirror_name, expires_at)
-       VALUES ($1, $2, $3, $4, $5, $6, $7, now() + interval '30 minutes')`,
-      [
-        id,
-        tokenHash(secret),
-        input.collection_id,
-        input.display_name,
-        input.source_name,
-        input.retain_mirror,
-        input.retain_mirror ? (input.mirror_name ?? input.source_name) : null
-      ]
-    );
-    return reply.code(201).send({
-      adoption_id: id,
-      adoption_secret: secret,
-      verification_uri: `${publicUrl}/adopt/${id}`,
-      expires_in: expiresIn
-    });
-  });
-
-  app.get("/v1/authority-adoptions/:adoptionId", async (request, reply) => {
-    const user = await requireUser(request, reply, options.db, options.tailscaleAuth);
-    if (!user) return;
-    const { adoptionId } = z.object({ adoptionId: z.uuid() }).parse(request.params);
-    await recoverExpiredAuthorityAdoptions(options.db, options.hostedProvider);
-    const found = await options.db.query<AuthorityAdoptionRow>(
-      `${authorityAdoptionSelect()}
-       WHERE adoption.id = $1
-         AND (adoption.expires_at > now() OR adoption.state NOT IN ('requested', 'approved', 'prepared'))
-         AND (adoption.user_id IS NULL OR adoption.user_id = $2)`,
-      [adoptionId, user.id]
-    );
-    const adoption = found.rows[0];
-    if (!adoption) {
-      return reply.code(404).send(apiError(
-        "authority_adoption_not_found",
-        "Collection adoption expired or was not found."
-      ));
-    }
-    return { adoption: authorityAdoptionView(adoption) };
-  });
-
-  app.post("/v1/authority-adoptions/:adoptionId/approve", async (request, reply) => {
-    if (!options.hostedCollections || !options.hostedProvider) {
-      return reply.code(404).send(apiError(
-        "hosted_adoption_unavailable",
-        "This Connect server cannot adopt local collections."
-      ));
-    }
-    const user = await requireUser(request, reply, options.db, options.tailscaleAuth);
-    if (!user) return;
-    const { adoptionId } = z.object({ adoptionId: z.uuid() }).parse(request.params);
-    z.object({}).strict().parse(request.body ?? {});
-    await recoverExpiredAuthorityAdoptions(options.db, options.hostedProvider);
-    const connection = await options.db.connect();
-    try {
-      await connection.query("BEGIN");
-      const locked = await connection.query<AuthorityAdoptionRow>(
-        `${authorityAdoptionSelect()}
-         WHERE adoption.id = $1 FOR UPDATE`,
-        [adoptionId]
-      );
-      const adoption = locked.rows[0];
-      if (
-        !adoption
-        || (adoption.user_id !== null && adoption.user_id !== user.id)
-        || new Date(adoption.expires_at).getTime() <= Date.now()
-      ) {
-        await connection.query("ROLLBACK");
-        return reply.code(404).send(apiError(
-          "authority_adoption_not_found",
-          "Collection adoption expired or was not found."
-        ));
-      }
-      if (adoption.state !== "requested") {
-        await connection.query("COMMIT");
-        if (["approved", "prepared", "activating", "completed"].includes(adoption.state)) {
-          return { adoption: authorityAdoptionView(adoption) };
-        }
-        return reply.code(409).send(apiError(
-          "authority_adoption_inactive",
-          "Collection adoption is no longer awaiting approval."
-        ));
-      }
-      const hosted = await connection.query<{ id: string }>(
-        `INSERT INTO hosted_collections
-           (id, user_id, display_name, template, provider_url, contracts,
-            authority_state, authority_epoch)
-         VALUES ($1, $2, $3, 'mdbase', $4, '[]'::jsonb, 'importing', $5)
-         ON CONFLICT (id) DO NOTHING
-         RETURNING id`,
-        [
-          adoption.collection_id,
-          user.id,
-          adoption.display_name,
-          options.hostedProvider.url,
-          Number(adoption.next_authority_epoch)
-        ]
-      );
-      if (!hosted.rows[0]) {
-        await connection.query("ROLLBACK");
-        return reply.code(409).send(apiError(
-          "authority_adoption_collection_conflict",
-          "A hosted collection already uses this collection identity."
-        ));
-      }
-      if (adoption.retain_mirror) {
-        await connection.query(
-          `INSERT INTO mirror_pairing_requests
-             (id, secret_hash, mirror_name, mode, user_id, collection_hint,
-              collection_id, approved_at, expires_at)
-           VALUES ($1, $2, $3, 'read_write', $4, $5, $5, now(), now() + interval '24 hours')`,
-          [
-            adoption.id,
-            adoption.secret_hash,
-            adoption.mirror_name ?? adoption.source_name,
-            user.id,
-            adoption.collection_id
-          ]
-        );
-      }
-      const approved = await connection.query<AuthorityAdoptionRow>(
-        `UPDATE authority_adoption_requests
-         SET user_id = $2, state = 'approved', approved_at = now()
-         WHERE id = $1 AND state = 'requested'
-         RETURNING *`,
-        [adoptionId, user.id]
-      );
-      await audit(connection, user.id, "authority_adoption.approved", adoptionId, {
-        collection_id: adoption.collection_id,
-        retain_mirror: adoption.retain_mirror
-      });
-      await connection.query("COMMIT");
-      return { adoption: authorityAdoptionView(approved.rows[0]) };
-    } catch (error) {
-      await connection.query("ROLLBACK");
-      throw error;
-    } finally {
-      connection.release();
-    }
-  });
-
-  app.post("/v1/authority-adoptions/:adoptionId/exchange", async (request, reply) => {
-    if (!options.hostedCollections || !options.hostedProvider) {
-      return reply.code(404).send(apiError(
-        "hosted_adoption_unavailable",
-        "This Connect server cannot adopt local collections."
-      ));
-    }
-    const { adoptionId } = z.object({ adoptionId: z.uuid() }).parse(request.params);
-    z.object({}).strict().parse(request.body ?? {});
-    const secret = bearerToken(request);
-    if (!secret) {
-      return reply.code(401).send(apiError(
-        "invalid_authority_adoption",
-        "Collection adoption credential required."
-      ));
-    }
-    await recoverExpiredAuthorityAdoptions(options.db, options.hostedProvider);
-    const adoption = await authorityAdoptionBySecret(options.db, adoptionId, secret);
-    if (!adoption) {
-      return reply.code(404).send(apiError(
-        "authority_adoption_not_found",
-        "Collection adoption expired or was not found."
-      ));
-    }
-    if (adoption.state === "requested") {
-      return reply.code(202).send({ status: "pending" });
-    }
-    if (adoption.state === "completed") {
-      return {
-        status: "completed",
-        adoption: authorityAdoptionView(adoption)
-      };
-    }
-    if (adoption.state === "activating") {
-      return {
-        status: "activating",
-        adoption: authorityAdoptionView(adoption)
-      };
-    }
-    if (adoption.state === "expired") {
-      return reply.code(409).send(apiError(
-        "authority_adoption_expired",
-        "Collection adoption expired before hosted activation began."
-      ));
-    }
-    if (adoption.state === "cancelled") {
-      return reply.code(409).send(apiError(
-        "authority_adoption_cancelled",
-        "Collection adoption was cancelled before hosted activation began."
-      ));
-    }
-    if (!["approved", "prepared"].includes(adoption.state)) {
-      return reply.code(409).send(apiError(
-        "authority_adoption_inactive",
-        "Collection adoption is no longer active."
-      ));
-    }
-    const importToken = randomToken("ati");
-    const prepared = await options.hostedProvider.prepareAuthorityImport({
-      transferId: adoption.id,
-      collectionId: adoption.collection_id,
-      displayName: adoption.display_name,
-      token: importToken,
-      authorityEpoch: Number(adoption.next_authority_epoch),
-      ttlSeconds: 30 * 60
-    });
-    const saved = await options.db.query<AuthorityAdoptionRow>(
-      `UPDATE authority_adoption_requests
-       SET state = 'prepared', prepared_at = COALESCE(prepared_at, now()),
-           expires_at = $2
-       WHERE id = $1 AND state IN ('approved', 'prepared')
-       RETURNING *`,
-      [adoption.id, prepared.expires_at]
-    );
-    if (!saved.rows[0]) {
-      return reply.code(409).send(apiError(
-        "authority_adoption_state_changed",
-        "Collection adoption changed state while upload access was prepared."
-      ));
-    }
-    return {
-      status: "ready",
-      adoption: authorityAdoptionView(saved.rows[0]),
-      staged: {
-        state: prepared.state,
-        manifest_digest: prepared.manifest_digest,
-        source_revision: prepared.source_revision,
-        source_head: prepared.source_head
-      },
-      import: authorityImportCapability(
-        options.hostedProvider.url,
-        adoption.id,
-        importToken
-      )
-    };
-  });
-
-  app.post("/v1/authority-adoptions/:adoptionId/complete", async (request, reply) => {
-    if (!options.hostedProvider) {
-      return reply.code(404).send(apiError(
-        "hosted_adoption_unavailable",
-        "This Connect server cannot adopt local collections."
-      ));
-    }
-    const { adoptionId } = z.object({ adoptionId: z.uuid() }).parse(request.params);
-    const secret = bearerToken(request);
-    if (!secret) {
-      return reply.code(401).send(apiError(
-        "invalid_authority_adoption",
-        "Collection adoption credential required."
-      ));
-    }
-    const input = z.object({
-      manifest_digest: z.string().regex(/^[a-f0-9]{64}$/),
-      source_revision: z.string().regex(/^sha256:[a-f0-9]{64}$/),
-      source_head: z.number().int().nonnegative()
-    }).strict().parse(request.body);
-    const adoption = await authorityAdoptionBySecret(options.db, adoptionId, secret);
-    if (!adoption) {
-      return reply.code(404).send(apiError(
-        "authority_adoption_not_found",
-        "Collection adoption was not found."
-      ));
-    }
-    if (adoption.state === "completed") {
-      if (
-        adoption.manifest_digest !== input.manifest_digest
-        || adoption.source_revision !== input.source_revision
-        || Number(adoption.final_head) !== input.source_head
-      ) {
-        return reply.code(409).send(apiError(
-          "authority_adoption_snapshot_mismatch",
-          "Completed adoption belongs to a different source snapshot."
-        ));
-      }
-      return {
-        status: "completed",
-        adoption: authorityAdoptionView(adoption)
-      };
-    }
-    if (
-      !["prepared", "activating"].includes(adoption.state)
-      || (
-        adoption.state === "prepared"
-        && new Date(adoption.expires_at).getTime() <= Date.now()
-      )
-    ) {
-      return reply.code(409).send(apiError(
-        "authority_adoption_inactive",
-        "Collection adoption is no longer prepared."
-      ));
-    }
-    if (
-      adoption.state === "activating"
-      && (
-        adoption.manifest_digest !== input.manifest_digest
-        || adoption.source_revision !== input.source_revision
-        || Number(adoption.final_head) !== input.source_head
-      )
-    ) {
-      return reply.code(409).send(apiError(
-        "authority_adoption_snapshot_mismatch",
-        "Authority activation must resume with the same fenced source snapshot."
-      ));
-    }
-    if (adoption.state === "prepared") {
-      const reserved = await options.db.query(
-        `UPDATE authority_adoption_requests
-         SET state = 'activating', manifest_digest = $2,
-             source_revision = $3, final_head = $4
-         WHERE id = $1 AND state = 'prepared'`,
-        [
-          adoption.id,
-          input.manifest_digest,
-          input.source_revision,
-          input.source_head
-        ]
-      );
-      if (reserved.rowCount !== 1) {
-        return reply.code(409).send(apiError(
-          "authority_adoption_state_changed",
-          "Collection adoption changed state while activation was reserved."
-        ));
-      }
-    }
-    const completed = await options.hostedProvider.completeAuthorityImport(
-      adoption.id,
-      input.manifest_digest,
-      input.source_revision
-    );
-    if (
-      completed.id !== adoption.id
-      || completed.collection_id !== adoption.collection_id
-      || completed.state !== "completed"
-      || completed.authority_epoch !== Number(adoption.next_authority_epoch)
-      || completed.manifest_digest !== input.manifest_digest
-      || completed.source_revision !== input.source_revision
-      || completed.source_head !== input.source_head
-    ) {
-      throw new RequestValidationError(
-        "The hosted authority activated a different adoption snapshot."
-      );
-    }
-    const connection = await options.db.connect();
-    try {
-      await connection.query("BEGIN");
-      const current = await connection.query<AuthorityAdoptionRow>(
-        `${authorityAdoptionSelect()}
-         WHERE adoption.id = $1 FOR UPDATE`,
-        [adoption.id]
-      );
-      if (current.rows[0]?.state === "completed") {
-        await connection.query("COMMIT");
-        return {
-          status: "completed",
-          adoption: authorityAdoptionView(current.rows[0])
-        };
-      }
-      if (current.rows[0]?.state !== "activating") {
-        throw new RequestValidationError(
-          "Collection adoption is not reserved for activation."
-        );
-      }
-      const activated = await connection.query(
-        `UPDATE hosted_collections
-         SET authority_state = 'active', authority_epoch = $2, contracts = $3::jsonb
-         WHERE id = $1 AND authority_state = 'importing'`,
-        [
-          adoption.collection_id,
-          completed.authority_epoch,
-          JSON.stringify(completed.contracts)
-        ]
-      );
-      const committed = await connection.query<AuthorityAdoptionRow>(
-        `UPDATE authority_adoption_requests
-         SET state = 'completed', contracts = $2::jsonb, completed_at = now()
-         WHERE id = $1 AND state = 'activating'
-         RETURNING *`,
-        [adoption.id, JSON.stringify(completed.contracts)]
-      );
-      if (activated.rowCount !== 1 || !committed.rows[0]) {
-        throw new RequestValidationError(
-          "Authority metadata changed while adoption completed."
-        );
-      }
-      await audit(connection, adoption.user_id!, "authority_adoption.completed", adoption.id, {
-        collection_id: adoption.collection_id,
-        authority_epoch: completed.authority_epoch,
-        retain_mirror: adoption.retain_mirror
-      });
-      await connection.query("COMMIT");
-      return {
-        status: "completed",
-        adoption: authorityAdoptionView(committed.rows[0])
-      };
-    } catch (error) {
-      await connection.query("ROLLBACK");
-      throw error;
-    } finally {
-      connection.release();
-    }
-  });
-
-  app.delete("/v1/authority-adoptions/:adoptionId", async (request, reply) => {
-    if (!options.hostedProvider) {
-      return reply.code(404).send(apiError(
-        "hosted_adoption_unavailable",
-        "This Connect server cannot adopt local collections."
-      ));
-    }
-    const { adoptionId } = z.object({ adoptionId: z.uuid() }).parse(request.params);
-    const secret = bearerToken(request);
-    if (!secret) {
-      return reply.code(401).send(apiError(
-        "invalid_authority_adoption",
-        "Collection adoption credential required."
-      ));
-    }
-    const adoption = await authorityAdoptionBySecret(options.db, adoptionId, secret);
-    if (!adoption) {
-      return reply.code(404).send(apiError(
-        "authority_adoption_not_found",
-        "Collection adoption was not found."
-      ));
-    }
-    if (adoption.state === "completed") {
-      return reply.code(409).send(apiError(
-        "authority_adoption_completed",
-        "Completed collection adoption cannot be cancelled."
-      ));
-    }
-    if (adoption.state === "activating") {
-      return reply.code(409).send(apiError(
-        "authority_adoption_activation_started",
-        "Hosted authority activation has started and can no longer be cancelled."
-      ));
-    }
-    if (adoption.state !== "cancelled") {
-      const cancelled = await options.db.query(
-        `UPDATE authority_adoption_requests
-         SET state = 'cancelled', cancelled_at = now()
-         WHERE id = $1 AND state IN ('requested', 'approved', 'prepared', 'expired')`,
-        [adoption.id]
-      );
-      if (cancelled.rowCount !== 1) {
-        return reply.code(409).send(apiError(
-          "authority_adoption_state_changed",
-          "Collection adoption changed state while cancellation was reserved."
-        ));
-      }
-    }
-    try {
-      await options.hostedProvider.abortAuthorityImport(adoption.id);
-    } catch (error) {
-      if (
-        !(error instanceof HostedProviderResponseError)
-        || !["authority_import_not_found", "authority_import_inactive"].includes(error.code)
-      ) {
-        throw error;
-      }
-    }
-    const connection = await options.db.connect();
-    try {
-      await connection.query("BEGIN");
-      await connection.query("DELETE FROM mirror_pairing_requests WHERE id = $1", [adoption.id]);
-      await connection.query(
-        `DELETE FROM hosted_collections
-         WHERE id = $1 AND authority_state = 'importing'`,
-        [adoption.collection_id]
-      );
-      if (adoption.user_id) {
-        await audit(connection, adoption.user_id, "authority_adoption.cancelled", adoption.id, {
-          collection_id: adoption.collection_id
-        });
-      }
-      await connection.query("COMMIT");
-      return { ok: true };
-    } catch (error) {
-      await connection.query("ROLLBACK");
-      throw error;
-    } finally {
-      connection.release();
-    }
+  registerAuthorityAdoptionRoutes(app, {
+    db: options.db,
+    publicUrl,
+    tailscaleAuth: options.tailscaleAuth,
+    hostedCollections: options.hostedCollections,
+    hostedProvider: options.hostedProvider
   });
 
   app.post("/v1/mirror-pairing-requests/:pairingId/authority-transfers", async (request, reply) => {
@@ -6049,111 +5533,6 @@ async function mirrorAuthorityTransfer(
   return access?.actions.has("authority.transfer") ? transfer : null;
 }
 
-function authorityAdoptionSelect(): string {
-  return `SELECT adoption.id, adoption.secret_hash, adoption.collection_id,
-                 adoption.display_name, adoption.source_name,
-                 adoption.retain_mirror, adoption.mirror_name,
-                 adoption.user_id, adoption.state,
-                 adoption.next_authority_epoch, adoption.final_head,
-                 adoption.manifest_digest, adoption.source_revision,
-                 adoption.contracts, adoption.expires_at,
-                 adoption.approved_at, adoption.prepared_at,
-                 adoption.completed_at
-          FROM authority_adoption_requests adoption`;
-}
-
-async function authorityAdoptionBySecret(
-  db: DatabasePool,
-  adoptionId: string,
-  secret: string
-): Promise<AuthorityAdoptionRow | null> {
-  const result = await db.query<AuthorityAdoptionRow>(
-    `${authorityAdoptionSelect()}
-     LEFT JOIN users account ON account.id = adoption.user_id
-     WHERE adoption.id = $1 AND adoption.secret_hash = $2
-       AND adoption.revoked_at IS NULL
-       AND (adoption.user_id IS NULL OR account.suspended_at IS NULL)`,
-    [adoptionId, tokenHash(secret)]
-  );
-  return result.rows[0] ?? null;
-}
-
-function authorityAdoptionView(adoption: AuthorityAdoptionRow): Record<string, unknown> {
-  return {
-    id: adoption.id,
-    collection_id: adoption.collection_id,
-    display_name: adoption.display_name,
-    source_name: adoption.source_name,
-    retain_mirror: adoption.retain_mirror,
-    mirror_name: adoption.mirror_name,
-    state: adoption.state,
-    authority_epoch: Number(adoption.next_authority_epoch),
-    final_head: adoption.final_head === null ? null : Number(adoption.final_head),
-    manifest_digest: adoption.manifest_digest,
-    source_revision: adoption.source_revision,
-    contracts: adoption.contracts ?? [],
-    expires_at: new Date(adoption.expires_at).toISOString(),
-    approved_at: adoption.approved_at
-      ? new Date(adoption.approved_at).toISOString()
-      : null,
-    prepared_at: adoption.prepared_at
-      ? new Date(adoption.prepared_at).toISOString()
-      : null,
-    completed_at: adoption.completed_at
-      ? new Date(adoption.completed_at).toISOString()
-      : null
-  };
-}
-
-async function recoverExpiredAuthorityAdoptions(
-  db: DatabasePool,
-  hostedProvider?: HostedProviderClient
-): Promise<void> {
-  await db.query(
-    `UPDATE authority_adoption_requests
-     SET state = 'expired'
-     WHERE expires_at <= now() AND state IN ('requested', 'approved', 'prepared')`
-  );
-  if (!hostedProvider) return;
-  const pendingCleanup = await db.query<{ id: string; collection_id: string }>(
-    `SELECT id, collection_id FROM authority_adoption_requests
-     WHERE state = 'expired' AND cleanup_completed = false`
-  );
-  for (const adoption of pendingCleanup.rows) {
-    try {
-      await hostedProvider.abortAuthorityImport(adoption.id);
-    } catch (error) {
-      if (
-        !(error instanceof HostedProviderResponseError)
-        || !["authority_import_not_found", "authority_import_inactive"].includes(error.code)
-      ) {
-        throw error;
-      }
-    }
-    const connection = await db.connect();
-    try {
-      await connection.query("BEGIN");
-      await connection.query("DELETE FROM mirror_pairing_requests WHERE id = $1", [adoption.id]);
-      await connection.query(
-        `DELETE FROM hosted_collections
-         WHERE id = $1 AND authority_state = 'importing'`,
-        [adoption.collection_id]
-      );
-      await connection.query(
-        `UPDATE authority_adoption_requests
-         SET cleanup_completed = true WHERE id = $1 AND state = 'expired'`,
-        [adoption.id]
-      );
-      await connection.query("COMMIT");
-    } catch (error) {
-      await connection.query("ROLLBACK");
-      throw error;
-    } finally {
-      connection.release();
-    }
-  }
-}
-
 async function recoverExpiredAuthorityTransfers(
   db: DatabasePool,
   hostedProvider?: HostedProviderClient,
@@ -6329,28 +5708,6 @@ function authorityImportTransferView(
     manifest_digest: transfer.manifest_digest,
     source_revision: transfer.source_revision,
     expires_at: new Date(transfer.expires_at).toISOString()
-  };
-}
-
-function authorityImportCapability(
-  providerUrl: string,
-  transferId: string,
-  accessToken: string
-): Record<string, string> {
-  const base = new URL(providerUrl);
-  const path = `/v1/authority-imports/${encodeURIComponent(transferId)}`;
-  const endpoint = (suffix: string) => {
-    const url = new URL(base);
-    url.pathname = `${path}/${suffix}`;
-    url.search = "";
-    url.hash = "";
-    return url.href;
-  };
-  return {
-    manifest_url: endpoint("manifest"),
-    records_url: endpoint("records"),
-    finalize_url: endpoint("finalize"),
-    access_token: accessToken
   };
 }
